pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obstask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-ai>=1.70.0",
    "pyyaml>=6.0",
    "questionary>=2.1.1",
    "rich>=14.3.3",
]
authors = [
    {name = "Juan Cobo Celdrán", email = "jcoboceldran@gmail.com"},
]
license = {text = "MIT"}
keywords = ["ai", "gemini", "cli", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

[tool.hatch.build.targets.wheel]
packages = ["obstask"]

[tool.hatch.build.targets.wheel.force-include]
"obstask/templates" = "obstask/templates"

[project.scripts]
ot = "obstask.cli:cli"
